requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
Django==4.0.4
django-extensions==3.1.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
executing==0.8.3
idna==3.3
ipdb==0.13.9
ipython==8.3.0
jedi==0.18.1
jsonfield==3.1.0
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
